# save-backend/backend-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SAVE Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5800",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/contests/active": {
      "get": {
        "description": "Get list of contests that are in progress now.",
        "operationId": "getContestsInProgress",
        "parameters": [
          {
            "description": "amount of contests that should be returned, default: 10",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Successfully fetched list of active contests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get list of contests that are in progress now.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/by-organization": {
      "get": {
        "description": "Get contests connected with given organization.",
        "operationId": "getOrganizationContests",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "query",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "amount of records that will be returned",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Successfully fetched public tests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Either contest with such name was not found or tests are not provided."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get contests connected with given organization.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/create": {
      "post": {
        "description": "Create a new contest.",
        "operationId": "createContest",
        "parameters": [
          {
            "description": "contest requested for creation",
            "in": "query",
            "name": "contestDto",
            "required": true
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Contest was successfully created."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "User cannot create contests with given organization."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Organization with given name was not found."
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Contest with given name is already present."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Create a new contest.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/featured/add-or-delete": {
      "post": {
        "description": "Mark contest to be featured if it is not marked so yet or unmark otherwise.",
        "operationId": "addOrDeleteContestToFeatured",
        "parameters": [
          {
            "in": "query",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "contest requested for creation",
            "in": "query",
            "name": "contestDto",
            "required": true
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Contest was successfully created."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Contest with given name is not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Create or delete featured contest.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/featured/list-active": {
      "get": {
        "description": "Get list of contests marked by admins as featured.",
        "operationId": "getFeaturedContests",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Contests were successfully fetched."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get featured contests.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/finished": {
      "get": {
        "description": "Get list of contests that has already finished.",
        "operationId": "getFinishedContests",
        "parameters": [
          {
            "description": "amount of contests that should be returned, default: 10",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Successfully fetched list of finished contests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get list of contests that has already finished.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/newest": {
      "get": {
        "description": "Get list of [pageSize] newest contests.",
        "operationId": "getSeveralNewestContests",
        "parameters": [
          {
            "description": "amount of records that will be returned",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Successfully fetched newest contests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContestDto"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get newest contests.",
        "tags": [
          "contests"
        ]
      }
    },
    "/api/v1/contests/update": {
      "post": {
        "description": "Change existing contest settings.",
        "operationId": "updateContest",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "query",
            "name": "contestRequest",
            "required": true
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successfully fetched public tests."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Not enough permission to edit current contest."
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Either organization or contest with such name was not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Update contest.",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/api/v1/contests/{contestName}": {
      "get": {
        "description": "Get contest by name.",
        "operationId": "getContestByName",
        "parameters": [
          {
            "description": "name of a contest",
            "in": "path",
            "name": "contestName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContestDto"
                }
              }
            },
            "description": "Successfully fetched contest by it\u0027s name."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContestDto"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContestDto"
                }
              }
            },
            "description": "Contest with such name was not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get contest by name.",
        "tags": [
          "contests"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContestDto": {
        "required": [
          "name",
          "organizationName",
          "status",
          "testSuiteIds"
        ],
        "type": "object",
        "properties": {
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "DELETED"
            ]
          },
          "testSuiteIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "description": "contest requested for creation"
      },
...
    },
    "securitySchemes": {
      "basic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  }
}